requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.0
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.26.0
qdrant-client==1.7.3
